Restructure configure command under cmd, remove public APIs
