+++ cmd/goreleaser/internal/configure.go
@@ -1,6 +1,18 @@
-//go:build releaser
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-package main
+package internal
 
 // This file is a script which generates the .goreleaser.yaml file for all
 // supported OpenTelemetry Collector distributions.
@@ -8,39 +20,18 @@ package main
 // Run it with `make generate-goreleaser`.
 
 import (
-	"flag"
 	"fmt"
-	"log"
-	"os"
 	"path"
 	"strings"
 
 	"github.com/goreleaser/goreleaser/pkg/config"
 	"github.com/goreleaser/nfpm/v2/files"
-	"gopkg.in/yaml.v2"
 )
 
 var (
 	ImagePrefixes = []string{"otel", "ghcr.io/open-telemetry/opentelemetry-collector-releases"}
 	Architectures = []string{"386", "amd64", "arm64", "ppc64le"}
-
-	distsFlag = flag.String("d", "", "Collector distributions(s) to build, comma-separated")
 )
-
-func main() {
-	flag.Parse()
-
-	if len(*distsFlag) == 0 {
-		log.Fatal("no distributions to build")
-	}
-	dists := strings.Split(*distsFlag, ",")
-
-	project := Generate(ImagePrefixes, dists)
-
-	if err := yaml.NewEncoder(os.Stdout).Encode(&project); err != nil {
-		log.Fatal(err)
-	}
-}
 
 func Generate(imagePrefixes []string, dists []string) config.Project {
 	return config.Project{
